+++ src/controllers/dataController.js
@@ -44,7 +44,6 @@ const findMatches = (input, query, ctx) => {
               mode: searchEngine,
               diacritics,
               highlight: resultItem.highlight,
-              className: (resultItem.highlight || {}).className,
             });
 
       if (!match) return;
@@ -76,6 +75,11 @@ const findMatches = (input, query, ctx) => {
 
   // Prepare data feedback object
   ctx.dataFeedback = { input, query, matches, results };
+
+  /**
+   * @emit {results} event on search response with matches
+   **/
+  eventEmitter("results", ctx);
 };
 
 export { getData, findMatches };